backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow>=10.0.0
tensorflow>=2.20.0
numpy==2.1.3
pydantic==2.5.3
python-dotenv==1.0.0
groq==0.4.1
httpx==0.26.0
aiofiles==23.2.1

